+++ test/utils.js
@@ -1,0 +1,26 @@
+define([
+
+  'lateralus'
+
+], function (
+
+  Lateralus
+
+) {
+  'use strict';
+
+  return {
+    /**
+     * @param {Function} [extraConstructorCode]
+     */
+    getLateraralusApp: function (extraConstructorCode) {
+      return Lateralus.beget(function () {
+        Lateralus.apply(this, arguments);
+
+        if (extraConstructorCode) {
+          extraConstructorCode.call(this);
+        }
+      });
+    }
+  };
+});