+++ changelog.md
@@ -1,3 +1,11 @@
+# 0.6.0
+
+  * Add mixins.amplify.
+  * Add Lateralus.prototype.dispose.
+  * Lateralus.Component#dispose triggers "beforeDispose" event.
+  * Extend Lateralus.Component.Model from Backbone.Model, not Lateralus.Model.
+  * Bug fixes.
+
 # 0.5.0
 
   * Add Lateralus.Router.